requirements.txt
django
django-filter
djangorestframework
djoser
psycopg2-binary